requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.2
spacy>=3.1.0
python-docx>=0.8.11
pdfminer.six>=20201018
nltk>=3.6.3
transformers>=4.11.0
torch>=1.9.0
Pillow>=8.0.0
pytesseract>=0.3.8
pdf2image>=1.16.0